encDef
