+++ src/consts.ts
@@ -31,11 +31,22 @@ export const TimeUnits = [
   'year', 'month', 'day', 'date', 'hours', 'minutes', 'seconds'
 ];
 
+// TODO: see if there is a nice way to import from Vega schema
+export const ValidAggregateOps = [
+  'values', 'count', 'valid', 'missing', 'distinct',
+  'sum', 'mean', 'average', 'variance', 'variancep', 'stdev',
+  'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max',
+  'argmin', 'argmax'
+];
+
 export const ENCODING_TYPES = [
   Enctype.X, Enctype.Y,
   Enctype.ROW, Enctype.COL,
   Enctype.SIZE, Enctype.SHAPE, Enctype.COLOR, Enctype.TEXT, Enctype.DETAIL
 ];
+
+// TODO: remove  
+export var MAXBINS_DEFAULT = 15;
 
 export const Shorthand = {
   Delim: '|',
